requirements.txt
altgraph==0.17.4
astroid==3.3.10
asttokens==3.0.0
bitarray==3.5.0
bitstring==4.3.1
cabarchive==0.2.4
cffi==1.17.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.5
cx_Freeze==8.4.0
cx_Logging==3.2.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
docutils==0.21.2
esptool==5.0.0
executing==2.2.0
filelock==3.19.1
fonttools==4.58.5
imbalanced-learn==0.13.0
intelhex==2.3.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lief==0.16.6
markdown-it-py==3.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy==1.16.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyamaze==1.0.1
pycparser==2.22
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pylint==3.3.7
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.0
reedsolo==1.7.0
rich==14.0.0
rich-click==1.8.9
scikit-learn==1.6.1
scipy==1.16.0
Send2Trash==1.8.3
setuptools==79.0.0
shiboken6==6.9.0
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
striprtf==0.0.29
thonny==4.1.7
threadpoolctl==3.6.0
tomlkit==0.13.2
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
wheel==0.45.1
